Fix integer comparison

+++ components/wear_levelling/WL_Ext_Perf.cpp
@@ -81,7 +81,7 @@ esp_err_t WL_Ext_Perf::erase_sector_fit(uint32_t start_sector, uint32_t count)
     uint32_t pre_check_start = start_sector % this->size_factor;
 
 
-    for (int i = 0; i < this->size_factor; i++) {
+    for (size_t i = 0; i < this->size_factor; i++) {
         if ((i < pre_check_start) || (i >= count + pre_check_start)) {
             result = this->read(start_sector / this->size_factor * this->flash_sector_size + i * this->fat_sector_size, &this->sector_buffer[i * this->fat_sector_size / sizeof(uint32_t)], this->fat_sector_size);
             WL_EXT_RESULT_CHECK(result);
@@ -91,7 +91,7 @@ esp_err_t WL_Ext_Perf::erase_sector_fit(uint32_t start_sector, uint32_t count)
     result = WL_Flash::erase_sector(start_sector / this->size_factor); // erase comlete flash sector
     WL_EXT_RESULT_CHECK(result);
     // And write back only data that should not be erased...
-    for (int i = 0; i < this->size_factor; i++) {
+    for (size_t i = 0; i < this->size_factor; i++) {
         if ((i < pre_check_start) || (i >= count + pre_check_start)) {
             result = this->write(start_sector / this->size_factor * this->flash_sector_size + i * this->fat_sector_size, &this->sector_buffer[i * this->fat_sector_size / sizeof(uint32_t)], this->fat_sector_size);
             WL_EXT_RESULT_CHECK(result);
